Time: 3 ms (68.00%) | Memory: 7.1 MB (18.45%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool judgeSquareSum(int c) {
        long long a=0;
-        long long b=sqrt(c);
+        int b=sqrt(c);
 
         while(a<=b){
         if(a*a + b*b ==c){
